+++ DrugSchedule.SqlServer/Data/Entities/Schedule/MedicamentTakingSchedule.cs
@@ -4,7 +4,7 @@ public class MedicamentTakingSchedule
 {
     public long Id { get; set; }
 
-    public required int UserProfileId { get; set; }
+    public required long UserProfileId { get; set; }
 
     public UserProfile? UserProfile { get; set; }
