+++ DrugSchedule.SqlServer/Data/Entities/UserData/UserProfileContact.cs
@@ -4,11 +4,11 @@ public class UserProfileContact
 {
     public long Id { get; set; }
 
-    public required int UserProfileId { get; set; }
+    public required long UserProfileId { get; set; }
 
     public UserProfile? UserProfile { get; set; }
 
-    public required int ContactProfileId { get; set; }
+    public required long ContactProfileId { get; set; }
 
     public UserProfile? ContactProfile { get; set; }
